GitCommit is now ParentGitCommit properly
also inside the bumper. Not sure now whether this information even matters but I suppose the tags match the tags. Parent git commit baked into the code does act as an authenticity audit chain, same as a blockchain, not entirely useless. It is just impossible to commit the hash inside the commit before making the commit.