Ripped out all the config/command stuff from p9
after looking at the old work again it was very clear I was reinventing the wheel in much of this codebase, now switching geers to take the previously done work and simplify it and get this thing done sooner. We want to use this as a framework in the future when we get the chance to focus on a wholly new system so it needs to be done sooner rather than later. It is something that can even help with marketing my work, since I have this toolset done, a lot of architecture problems are non issues.